requirements.txt
meshcore>=2.2.3
rrdtool-bindings>=0.5.0
pyserial>=3.5
jinja2>=3.1.0
